# server/address-service.openapi.yaml
openapi: 3.0.3
info:
  title: AddressService API
  description: This is the address-service API of InfinityApp by jonnef
  version: 1.0.0
servers:
  - url: http://infinityapp-api.com
  - url: http://jonnef-api.com
paths:
  /addresses:
    description: Address Resource
    get:
      tags:
      - addresses
      description: Get a list of all addresses
      responses:
        200:
          description: Success!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/addresses'
    post:
      tags:
      - addresses
      description: Add a new Person
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addressInfo'
      responses:
        201:
          description: Successfully added
  /addresses/{id}:
    description: Address methods with path parameter id
    get:
      tags:
      - addresses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success Response with Path Parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/address'
    patch:
      tags:
      - addresses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addressInfo'
      responses:
        200:
          description: "Successfully changed data of address with id: {id}"
    delete:
      tags:
      - addresses
      description: Delete Address by ADID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Successfully deleted the Address
components:
  schemas:
    address:
      type: object
      allOf:
        - $ref: '#/components/schemas/baseEntity'
        - $ref: '#/components/schemas/addressInfo'
    addressInfo:
      type: object
      properties:
        street:
          type: string
          example: Duisburgerstraße
        houseNumber:
          type: string
          example: 145
        postCode:
          type: integer
          example: 40477
        city:
          type: string
          example: Düsseldorf
        state:
          type: string
          example: North Rhine-Westphalia
    addresses:
     type: array
     items:
       allOf:
         - $ref: '#/components/schemas/address'
    baseEntity:
      properties:
        id:
          type: integer
          nullable: false
          example: 0
      required:
        - id